added test cases for 284,  301, 376, 379, 427, 438, 457, 640, 653

+++ no_loop_problems/284/specification.h
@@ -21,4 +21,4 @@
     complete behaviors;
     disjoint behaviors;
     */
-void calculateSharedTime(int l1, int r1, int l2, int r2, int k, int *out);
+void calculateSharedTime(long l1, long r1, long l2, long r2, long k, long *out);
