+++ no_loop_problems/301/solution.c
@@ -2,14 +2,14 @@
     There are three friend living on the straight line Ox in Lineland. The first friend lives at the point x 1 , the second friend lives at the point x 2 , and the third friend lives at the point x 3 . They plan to celebrate the New Year together, so they need to meet at one point. What is the minimum total distance they have to travel in order to meet at some point and celebrate the New Year? It's guaranteed that the optimal answer is always integer.
 */
 
-/*@ predicate IsValidSolution(long x1, long x2, long x3, long result) =
-    \exists long point;
+/*@ predicate IsValidSolution(int x1, int x2, int x3, int result) =
+    \exists int point;
     0 <= point &&
     \abs(point - x1) + \abs(point - x2) + \abs(point - x3) == result;
 */
 
-/*@ predicate ExistsSmallerSolution(long x1, long x2, long x3, long result) =
-    \exists long smaller_solution;
+/*@ predicate ExistsSmallerSolution(int x1, int x2, int x3, int result) =
+    \exists int smaller_solution;
     IsValidSolution(x1, x2, x3, smaller_solution) &&
     smaller_solution < result;
 */
@@ -24,10 +24,10 @@
     ensures IsValidSolution(x1, x2, x3, *out);
     ensures !ExistsSmallerSolution(x1, x2, x3, *out);
 */
-
-void calculateOptimalMeetingPointDistance(long x1, long x2, long x3, long *out)
+void calculateOptimalMeetingPointDistance(int x1, int x2, int x3, int *out)
 {
-    long max = x1;
+
+    int max = x1;
     long min = x1;
     if (x2 > max)
         max = x2;
